Added a bad way to compensate for latency
I saw an issue while not running it on my local, and decided that
I need to fix it. So, the server runs at a 60 fps update cycle, anything
in between those updates just returns the same thing as before with nothing unchanged.
however players can change direction between frames, so that may be an issue sooner or later.